Refactored schema controller, renamed schemas to schema
